20240314:ChamindaJ: use variable names consistent with scouting app csv file
we had multiple redundant variables that could not have been used based on the current scouting app output.  eg: we don't have any way to get the number of amplified  vs non-amplified amp or speaker shots with the csv file.
Also updated variable names to be consistent with the names used in the csv file